backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
httpx
google-generativeai
